requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
SQLAlchemy==2.0.13
psycopg2-binary==2.9.6
bcrypt==4.0.1
python-jose==3.3.0
python-dotenv==1.0.0
python-multipart==0.0.6
pydantic[email]